Harden AR testing config
Without column type hints, ActiveRecord fixtures are loaded with generic
quoting.  Ruby strings, for example, are assumed to be destined for plain
string columns.  For blob columns, special quoting is needed and this
approach is inadequate.  To provide proper column type hints, it is
critical that Rails know the ActiveRecord class of the fixtures being
loaded -a table name is inadequate (at least for namespaced model files).
The solution is to use the #set_fixture_class method.