Make Sources::File#public_uri return nil
If a public URI is both available and desired for file sources, the client application should
generate it.  In principle, the plugin has no knowledge of the access control requirements of
the client application.  Furthermore, the specious comparison of the file store to the Rails
public path assumes that application servers are asset hosts.